Creada clase Visualizador para modularizar más el controlador
