+++ Genetic Circuit/src/vista/Vista.java
@@ -3,7 +3,6 @@ package vista;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.swing.SwingUtilities;
 import javax.swing.UIManager;
 import javax.swing.UnsupportedLookAndFeelException;
